+++ src/path.js
@@ -151,8 +151,17 @@ const Path = {
       if (s.type === PATH_SEGMENT_TYPES.Separator) {
         return Path.PathSegment.String("/");
       } else if (s.type === PATH_SEGMENT_TYPES.STRING) {
+        // This allows for Path segments [(STRING, foofilename), (STRING, .bar)]
+        // but forbids [(SEPARATOR, /), (STRING, .bar)]
+        // STRING followed by a STRING can happen in filename rewrites
+        const forbidLeadingDots =
+          i === 0 || buf[i - 1].type === PATH_SEGMENT_TYPES.SEPARATOR;
         return Path.PathSegment.String(
-          Path.sanitizeFilename(s.val, options.truncateLength)
+          Path.sanitizeFilename(
+            s.val,
+            options.truncateLength,
+            forbidLeadingDots
+          )
         );
       } else {
         return s;